Because I encrypted admin passwords in the db this can no longer do simple login authentication.  It now makes http call to mathspring to authenticate the user/pass .
I also added a util/config.js file to put in config settings.   This has config.pageContext so that I don't have to use the -c program argument anymore.
It does mean having a different copy of this file on the server and putting it in the gitignore so that doing pulls on the server doesn't overwrite what is there.